feedproducer: WIP unfinished file-based repo impls
I think it'd be a neat trick, but the fundamental difficulty with file-based repos is updating the lookup indices. For entities in particular, when updating we have to all entity IDs to the index for their new pages, then save the actual updates, then remove them from the index for their old pages. That's all doable, but it's also a pretty straightforward index update problem that I don't think is interesting to show off.